+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-14](https://github.com/ITK-Leantime/leantime-dataexport/pull/15)
+  Set timezone on returned values from getDateTime to utc
 * [PR-14](https://github.com/ITK-Leantime/leantime-dataexport/pull/14)
   Cleaned up code checks and release script
 * [PR-13](https://github.com/ITK-Leantime/leantime-dataexport/pull/13)